DSR-294: rename _Page to _BasicPage
This name became too generic for its purpose, which is
a mixin for the basic pages of content.
