+++ pages/_lang/index.vue
@@ -33,7 +33,7 @@
 <script>
   // Mixins
   import Global from '~/components/_Global';
-  import Page from '~/components/_Page';
+  import BasicPage from '~/components/_BasicPage';
 
   // Components
   import AppBar from '~/components/AppBar';
@@ -49,7 +49,7 @@
   export default {
     mixins: [
       Global,
-      Page,
+      BasicPage,
     ],
 
     // Declare any components we're using here
